requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings

langchain_community
langchain
langchain-google-genai
langchain-pinecone
pypdf
lark

requests

google-generativeai==0.7.2
pymupdf
pinecone-client
pinecone
tqdm